requirements.txt
# ─────────────────────────────────────────────────────────
#  Job AI Agent FREE — All tools are completely free to use
# ─────────────────────────────────────────────────────────

# AI Engine — Groq API (free tier, no credit card needed)
groq>=0.9.0

# Web scraping
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=5.2.1

# Browser automation (optional — only needed for JS-heavy sites)
playwright>=1.44.0

# Web dashboard API
fastapi>=0.111.0
uvicorn>=0.30.0

# Email + utilities
python-dotenv>=1.0.1
PyYAML>=6.0.1
schedule>=1.2.1
rich>=13.7.1
